Create new "Zoom to 100%" action
Summary:
This patch implements a "Zoom to 100%" action and sticks it in the {nav View} menu. Since it's a `KStandardAction` with a `KStandardShortcut`, we automatically get the correct icon and keyboard shortcut, but we do override the name to be "Zoom to 100%" since that's clearer for Okular's use case.

FEATURE: 400048
FIXED-IN: 18.12.0

Test Plan:
- Action works to zoom the document to 100% scale when invoked
- Action is disabled when document is opened at 100% scale or is manually zoomed to 100% scale after being opened
- All other zoom modes and action still work

{F6341045}

Reviewers: #okular, #vdg, abetts

Reviewed By: #vdg, abetts

Subscribers: davidhurka, sander, tobiasdeiminger, veqz, abetts, aacid, okular-devel

Tags: #okular

Differential Revision: https://phabricator.kde.org/D16345