Regularize tool names
Summary:
Currently, some of the items in the {nav Tools} menu have "tool" in the name, and some
do not. This is in any event redundant since they're in a menu with that word in it
already. In the toolbar, these actions become more inconsistent since their names change
and can even differ between what's shown in the pop-up menu and what's shown on the
button text.

This patch regularizes the names, stops changing the names for the toolbar versions, and
updates the docbook accordingly.

Test Plan:
{F6874459, size=full}

{F6874458, size=full}

Reviewers: #okular, #vdg, filipf

Reviewed By: #vdg, filipf

Subscribers: filipf, okular-devel, kde-doc-english

Tags: #okular, #documentation

Differential Revision: https://phabricator.kde.org/D21621