Add "Open Containing Folder" feature
Summary:
This patch adds an "Open Containing Folder" menu item to the file menu, similar to
other KDE apps like Gwenview that have it there. This action is especially useful
for the case when you've downloaded a PDF from the internet that opens itself in
Okular. The location of this file may not be clear or easy to find without this
feature, and on several occasiona I have found myself wishing for it when this
happens.

Test Plan:
- Delete or move aside `~/.local/config/kxmlgui5/okular/part.rc`
- Menu item is enabled and works when there is an open document: {F6940449, size=full}
- Menu item is disabled when there is no open document: {F6940448, size=full}

Reviewers: #okular, aacid, sander

Reviewed By: #okular, aacid

Subscribers: shubham, davidhurka, okular-devel

Tags: #okular

Differential Revision: https://phabricator.kde.org/D22145