Do not use slotChangeDrawingToolEngine to reset the drawingEngine
The previous code used two consecutive calls to the method
slotChangeDrawingToolEngine to reset the m_drawingEngine
member (i.e., to delete it and to create a new one from scratch).
That worked; however, as a side effect the calls to
slotChangeDrawingToolEngine also set the cursor to an arrow,
which happens every time one continuous stroke is finished.
This went unnoticed because when drawing in presentation mode
the cursor is an arrow anyway.  However, in anticipation
of future patches that change the cursor to a cross-hair
when drawing, this explicit setting of the arrow shape
needs to go.  Therefore this patch replaces the two calls
to slotChangeDrawingToolEngine by the code they contain,
minus the code to set the cursor shape. The new code is
only one line longer, and more readable too.