"Fix" PartTest::testAdditionalActionTriggers
By waiting for the delayResizeEventTimer to trigger before starting the
test.

The test "randomly" failed before, which shows there's an actual bug in
the code, but the actual bug is "if you are moving the mouse over a
field that has a on focus action and at that very same time the
delayResizeEventTimer triggers, the action won't be executed".

That's almost impossible to reproduce in real life so i'm going to
pretend the bug is not there for all purposes