Move Continuous option to View Mode submenu
Summary:
This moves the Continuous option from the View menu to the
View Mode submenu. This makes the View Mode menu (a KActionMenu)
more useful in the main toolbar. Additionally, “Continuous” is explained by the context “View Mode”.

The primary intention was to give this View Mode a similar usage pattern like
the Change Colors menu (D21195), if both are added to the toolbar. See my
suggestion in Bugs 407217 and 407326.

FEATURE: 407326

Screenshot before:
{F6821917}

Screenshot now:
{F6821920}
and in the toolbar:
{F6821921}

Test Plan:
* Look into View menu and test entries
* Add View Mode menu to toolbar and test entries

Reviewers: #okular, #vdg, aacid

Spies: aacid, ngraham, okular-devel, kde-doc-english

Tags: #okular, #documentation

Differential Revision: https://phabricator.kde.org/D21196