XPS: fix multipiece image loading
Turns out some files contain images saved in the xps archive as a directory containing multiple pieces, the first one being the image itself and the second one being empty.
Of course this is quite a weird way to save an image, but it seems that it still is valid and other viewers are able to load it correctly.
This change fixes the loading of multipiece images.

(cherry picked from commit 3506fb2ea8bb052cc6e166e613370a1517248b37)