Add missing clang-format commit to .git-blame-ignore-revs
After 79514495 reformatted Okular's codebase with `clang-format`, that
commit was marked to be ignored by `git blame` by adding it to
`.git-blame-ignore-revs` in c907d485.

Later, !296 was merged but still needed reformatting, which was done in
fd244ecb. However, the latter commit was still missing from
`.git-blame-ignore-revs`, so let's add it.

Test Plan:
All commits mentioning `clang-format` which reformatted the codebase are
now included in `.git-blame-ignore-revs`.