Pass Android content: URLs fully encoded to Android API
This fixes opening files e.g. with a space anywhere in their name or path.

content: URLs we get from the file dialog need to get back into Android
API (via QFile or otherwise) as exactly the same string, any kind of
changes e.g. to the URL encoding are not considered to be the same file,
and thus a file we have no permission to access.