No need to lookup paths for images without a dot.
if there is no suffix, the image reader in Qt cycles thru all of the
possible values that it might be able to load to see if such a thing
exists relative to current workdir. If one has configured an annotation
with path from a image, it is very unlikely that such a image does not
have a suffix.

While this is not a problem for most users, on certain so-called
enterprise systems with encrypted file systems and virus scanners and so
on, it makes a slight difference.

(cherry picked from commit 0018d4fd117b62b33918b8ff3897adf8b058416b)