Fix DocumentHasPassword check
A document can be encrypted without requiring a password to open/sign it.

Currently we assume we need a password to sign it for such documents, but
that's not correct.

Instead of checking whether the document is encrypted remember whether
we needed a password for opening.

BUG: 474888
(cherry picked from commit 377632d689a21c5ee41c0c18691b1cf56d39ec5d)