Fix numerical interpretation in form fields
With the previous implementations of AFNumber_Keystroke [MR 988](https://invent.kde.org/graphics/okular/-/merge_requests/988), we ensure that during a keystroke event, a user does not enter thousands separator. With this in mind, numerical fields can always be first converted to en_US format and then used for calculations. This ensures that all numbers are correctly interpreted.

BUG: 474661