Fix the order of execution of events for text form fields
It is not required to recalculate the form fields during input. Only when the form fields are committed, do we need to recalculate the forms.
It is required to execute keystroke commit, validate and calculate events when form fields are set through javascript. The multiline text fields, do not have the FocusIn, FocusOut and Validate events execute. So added support for that as well.

All of the above changes have been made only for FormLineEdit and TextAreaEdit for now.