Fix build without QML
QTP0001 is declared in Qt Declarative and is only needed for the QML
version, so put it in the if BUILD_MOBILE block.

The error was:
QTP0001 is not a known Qt policy.  Did you include the necessary Qt module?


(cherry picked from commit 2efb01f13f3ff17ff36b849bf7b88e62052ed4df)

Co-authored-by: Carl Schwan <carl.schwan@gnupg.com>