Use overprint preview for PDF printers too
Since we're using PS as pivot format and convert back to PDF only later,
the overprint information isn't channeled properly all the way to the
printer. This result in confusion with all the overprint elements being
gone. So instead we activate the overprint preview if that's what was
displayed to have a consistent result on screen and while printing.